Released

+++ PetShop.Model/Employee.cs
@@ -11,7 +11,7 @@ namespace PetShop.Model
         public string Name { get; set; }
         public string Surname { get; set; }
         public EmployeeType EmployeeType { get; set; }
-        public double SallaryPerMonth { get; set; }
+        public decimal SallaryPerMonth { get; set; }
         //public decimal SallaryPerMonth { get; set; }
 
         //Relations
